requirements.txt
altgraph==0.17.5
beautifulsoup4==4.14.3
blinker==1.9.0
bottle==0.13.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
deep-translator==1.11.4
Flask==3.1.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
macholib==1.16.4
MarkupSafe==3.0.3
packaging==26.0
proxy-tools==0.1.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Security==12.0
pyobjc-framework-UniformTypeIdentifiers==12.0
pyobjc-framework-WebKit==12.0
pywebio==1.8.4
pywebview==6.1
requests==2.32.5
soupsieve==2.8.3
tornado==6.5.4
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.6.3
user-agents==2.2.0
Werkzeug==3.1.5
zipp==3.23.0
